requirements.txt
torch>=2.3.0
torchvision>=0.18.0
git+https://github.com/facebookresearch/sam2
omegaconf>=2.3.0
numpy>=1.26.0
opencv-python>=4.9.0
Pillow>=10.3.0
albumentations>=1.4.0
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
owslib>=0.29.0
scikit-learn>=1.4.0
tensorboard>=2.16.0
tqdm>=4.66.0
pyproj>=3.6.0
